Huge amount of changes:
* New classes: Moveable, Role, Status, Type, Node, XMLParser

Parsing XML files should now be easier and more reproduceable

Data is slowly being cleaned up, hence the addition of Status, Role, and
Type. 

Changed 'Object' to 'Model'

External data is now stored in library files: EntityLib.xml,
BuildingLib.xml, ItemLib.xml, ShaderLib.xml. 

All calls to the javax Vector2d or Vector3d classes have been replaced
with calls to the lwjgl Vector2f and Vector3f classes.

+++ src/com/cube/core/Building.java
@@ -3,10 +3,8 @@ package com.cube.core;
 import java.util.ArrayList;
 import java.util.HashMap;
 
-import javax.vecmath.Vector3d;
-
+import org.lwjgl.util.vector.Vector3f;
 import org.lwjgl.opengl.GL11;
-
 import com.cube.util.Utilities;
 
 public class Building extends Entity {
@@ -154,7 +152,7 @@ public class Building extends Entity {
 		
 		for (Entity e : Resources.entities) {
 			
-			if (Physics.distSquared(new Vector3d(pos[0], pos[1], pos[2]), e.position) < 1) {
+			if (Physics.distSquared(new Vector3f(pos[0], pos[1], pos[2]), e.position) < 1) {
 				
 			}
 		}
