+++ src/com/cube/states/FleeState.java
@@ -1,13 +1,13 @@
 package com.cube.states;
 
-import javax.vecmath.Vector2d;
+import org.lwjgl.util.vector.Vector2f;
 
 import com.cube.core.Entity;
 import com.cube.core.Physics;
 
 public class FleeState extends State {
 	
-	Vector2d tempVect;
+	Vector2f tempVect;
 	
 	@Override
 	public void enter(Entity e) {
